[2023.2] Fix shader warning in AtmosphericScattering
Fix shader warning due to backport

+++ Packages/com.unity.render-pipelines.high-definition/Runtime/Sky/PhysicallyBasedSky/ShaderVariablesPhysicallyBasedSky.cs.hlsl
@@ -25,7 +25,7 @@ GLOBAL_CBUFFER_START(ShaderVariablesPhysicallyBasedSky, b2)
     float _AtmosphericRadius;
     float _AerosolAnisotropy;
     float _AerosolPhasePartConstant;
-    float _CelestialLightCount;
+    uint _CelestialLightCount;
     float _AirDensityFalloff;
     float _AirScaleHeight;
     float _AerosolDensityFalloff;
